statistic_microservice/pyproject.toml
[tool.poetry]
name = "poetry-project"
version = "0.1.0"
description = ""
authors = ["iMaksus <max1112vas@gmail.com>"]
readme = "README.md"
packages = [{include = "poetry_project"}]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.85.1"
uvicorn = "^0.19.0"
boto3 = "^1.25.0"
pydantic = "^1.10.2"
python-dotenv = "^0.21.0"
botocore = "^1.28.1"
awscli = "^1.26.2"
localstack = "^1.2.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
